requirements.txt
openai-whisper
sounddevice
scipy
flask
soundfile
ffmpeg-python